Modify copySYCLKernelAttrs signature to take reference to KernelCallOperator
